test(session): gate TmuxSurvivesLoginSessionRemoval behind hostsensitive tag (#969)
This systemd user-scope cgroup test depends on:
  - systemd-run --user being available
  - a running user systemd manager that tracks MainPID for the spawned
    scope (NOT the case on most CI runners, inside nested tmux, or hosts
    without `loginctl enable-linger`)
  - host having no lingering tmux sessions racing the scope teardown

Moved the test plus its two test-only helpers
(startFakeLoginScope, startAgentDeckTmuxInUserScope) behind the
hostsensitive build tag so default pre-push / CI runs stay deterministic.
Opt in via:

    go test -tags hostsensitive -race ./internal/session/...

Refs: #969

Committed by Ashesh Goplani
